Replace extensions of Array and Optional with empty state and RoomListState.
