+++ Chatkit/Redux/State/MasterState.swift
@@ -4,14 +4,17 @@ struct MasterState: State {
     // MARK: - Properties
     
     let users: [UserState]
-    let currentUser: UserState?
+    let currentUser: UserState
     
-    let joinedRooms: [RoomState]
+//    let joinedRooms: [String : RoomState]
+    let joinedRooms: RoomListState
+//    let readStates: [String : ReadState]
     
     // TODO: Add readStates.
     // TODO: Add memberships.
     
-    static let empty: MasterState = MasterState(users: [], currentUser: nil, joinedRooms: [])
+//    static let empty: MasterState = MasterState(users: [], currentUser: nil, joinedRooms: [:], readStates: [:])
+    static let empty: MasterState = MasterState(users: [], currentUser: .empty, joinedRooms: .empty)
     
 }
