+++ Chatkit/Redux/State/RoomListState.swift
@@ -1,0 +1,14 @@
+
+struct RoomListState: State {
+    
+    // MARK: - Properties
+    
+    let rooms: [String : RoomState]
+    
+    static let empty: RoomListState = RoomListState(rooms: [:])
+    
+}
+
+// MARK: - Equatable
+
+extension RoomListState: Equatable {}
