+++ Chatkit/Redux/State/UserState.swift
@@ -1,6 +1,7 @@
 
 enum UserState: State {
     
+    case empty
     // TODO: Consider in future for DependencyFetcher
 //    case partial(identifier: String)
     case populated(identifier: String, name: String)
@@ -12,5 +13,3 @@ enum UserState: State {
 // MARK: - Equatable
 
 extension UserState: Equatable {}
-
-extension Optional: State where Wrapped == UserState {}
